Extract expand/collapse animation for reuse
